requirements.txt
supervisely==6.17.4
ffmpeg_python>=0.2.0
